refactor: 초기 기물 배치를 초기화 하는 방식 변경
